feat: add Preferences tab and implement auto-mark as read functionality
