+++ app/pages/(app)/settings/preferences.vue
@@ -1,0 +1,68 @@
+<script setup lang="ts">
+import { useAutoMarkSeenStore } from '~/composables/stores/useAutoMarkSeenStore';
+
+useSeoMeta({
+	title: 'Preferences | Delivr',
+	description: 'Manage your Delivr preferences',
+});
+
+// ── Auto-mark-as-read (global preference) ──
+const autoMarkSeenStore = useAutoMarkSeenStore();
+await autoMarkSeenStore.use();
+const autoMarkSeen = computed({
+	get: () => autoMarkSeenStore.enabled.value,
+	set: (value: boolean) => { autoMarkSeenStore.update({ enabled: value }); },
+});
+</script>
+
+<template>
+	<UDashboardPanel id="settings" :ui="{ body: 'lg:py-12 lg:gap-12 w-full lg:max-w-3xl mx-auto' }">
+		<template #header>
+			<DashboardPageHeader
+				title="Preferences"
+				icon="i-lucide-sliders-horizontal"
+				description="Control how Delivr behaves"
+			/>
+		</template>
+
+		<template #body>
+			<DashboardPageBody>
+				<!-- Header -->
+				<div>
+					<h2 class="text-xl font-semibold text-white">Preferences</h2>
+					<p class="text-sm text-slate-400 mt-1">Fine-tune how mail and the app behave</p>
+				</div>
+
+				<!-- Mail Preferences Card -->
+				<div class="rounded-xl border border-slate-800 bg-slate-900/60 backdrop-blur-sm overflow-hidden">
+					<div class="px-6 py-4 border-b border-slate-800">
+						<div class="flex items-center gap-3">
+							<div class="w-10 h-10 rounded-lg bg-sky-500/10 flex items-center justify-center">
+								<UIcon name="i-lucide-mail-open" class="w-5 h-5 text-sky-400" />
+							</div>
+							<div>
+								<h3 class="font-medium text-white">Mail Preferences</h3>
+								<p class="text-sm text-slate-400">Control how reading mail behaves</p>
+							</div>
+						</div>
+					</div>
+
+					<div class="p-6">
+						<UFormField
+							name="autoMarkSeen"
+							label="Auto-mark as read"
+							description="Automatically mark an email as read shortly after you open it."
+							class="flex max-sm:flex-col justify-between items-start gap-4"
+							:ui='{
+								root: "w-full sm:w-auto",
+								container: "w-full sm:w-auto",
+							}'
+						>
+							<UCheckbox v-model="autoMarkSeen" />
+						</UFormField>
+					</div>
+				</div>
+			</DashboardPageBody>
+		</template>
+	</UDashboardPanel>
+</template>
